@@ -24,9 +24,9 @@
     "babel": "^5.8.23",
     "express": "^4.12.4",
     "musicmetadata": "^2.0.0",
-    "q": "^1.4.1",
-    "react": "^0.13.3",
-    "reflux": "^0.2.12",
+    "react": "^0.14.0",
+    "react-dom": "^0.14.0",
+    "reflux": "^0.3.0",
     "reqwest": "^2.0.4"
   },
   "browserify": {
@@ -38,6 +38,7 @@
     "babelify": "^6.3.0",
     "browserify": "^11.2.0",
     "gulp": "^3.9.0",
+    "gulp-autoprefixer": "^3.1.0",
     "gulp-less": "^3.0.3",
     "gulp-uglify": "^1.4.1",
     "vinyl-buffer": "^1.0.0",
